Introduce file I/O
- Update LapBear.js to write results to .txt file
- Add start/end race buttons to home page (TODO: remove later)
